+++ src/apphosting/rollout.spec.ts
@@ -22,8 +22,7 @@ describe("apphosting rollouts", () => {
   const gitRepoLinkId = `${user}-${repo}`;
   const buildAndRolloutId = "build-2024-10-01-001";
 
-  let getBackendForLocationStub: sinon.SinonStub;
-  let getBackendForAmbiguousLocationStub: sinon.SinonStub;
+  let getBackend: sinon.SinonStub;
   let getRepoDetailsFromBackendStub: sinon.SinonStub;
   let listAllBranchesStub: sinon.SinonStub;
   let getGitHubBranchStub: sinon.SinonStub;
@@ -36,12 +35,7 @@ describe("apphosting rollouts", () => {
   let sleepStub: sinon.SinonStub;
 
   beforeEach(() => {
-    getBackendForLocationStub = sinon
-      .stub(backend, "getBackendForLocation")
-      .throws("unexpected getBackendForLocation call");
-    getBackendForAmbiguousLocationStub = sinon
-      .stub(backend, "getBackendForAmbiguousLocation")
-      .throws("unexpected getBackendForAmbiguousLocation call");
+    getBackend = sinon.stub(backend, "getBackend").throws("unexpected getBackend call");
     getRepoDetailsFromBackendStub = sinon
       .stub(devConnect, "getRepoDetailsFromBackend")
       .throws("unexpected getRepoDetailsFromBackend call");
@@ -149,8 +143,7 @@ describe("apphosting rollouts", () => {
 
     describe("createRollout", () => {
       it("should create a new rollout from user-specified branch", async () => {
-        getBackendForLocationStub.resolves(backend);
-        getBackendForAmbiguousLocationStub.resolves(backend);
+        getBackend.resolves(backend);
         getRepoDetailsFromBackendStub.resolves(repoLinkDetails);
         listAllBranchesStub.resolves(branches);
         getGitHubBranchStub.resolves(branchInfo);
@@ -160,7 +153,7 @@ describe("apphosting rollouts", () => {
         pollOperationStub.onFirstCall().resolves(rollout);
         pollOperationStub.onSecondCall().resolves(build);
 
-        await createRollout(backendId, projectId, location, branchId, undefined, true);
+        await createRollout(backendId, projectId, branchId, undefined, true);
 
         expect(createBuildStub).to.be.called;
         expect(createRolloutStub).to.be.called;
@@ -168,8 +161,7 @@ describe("apphosting rollouts", () => {
       });
 
       it("should create a new rollout from user-specified commit", async () => {
-        getBackendForLocationStub.resolves(backend);
-        getBackendForAmbiguousLocationStub.resolves(backend);
+        getBackend.resolves(backend);
         getRepoDetailsFromBackendStub.resolves(repoLinkDetails);
         getGitHubCommitStub.resolves(commitInfo);
         getNextRolloutIdStub.resolves(buildAndRolloutId);
@@ -178,7 +170,7 @@ describe("apphosting rollouts", () => {
         pollOperationStub.onFirstCall().resolves(rollout);
         pollOperationStub.onSecondCall().resolves(build);
 
-        await createRollout(backendId, projectId, location, undefined, commitSha, true);
+        await createRollout(backendId, projectId, undefined, commitSha, true);
 
         expect(createBuildStub).to.be.called;
         expect(createRolloutStub).to.be.called;
@@ -186,8 +178,7 @@ describe("apphosting rollouts", () => {
       });
 
       it("should prompt user for a branch if branch or commit ID is not specified", async () => {
-        getBackendForLocationStub.resolves(backend);
-        getBackendForAmbiguousLocationStub.resolves(backend);
+        getBackend.resolves(backend);
         getRepoDetailsFromBackendStub.resolves(repoLinkDetails);
         promptGitHubBranchStub.resolves(branchId);
         getGitHubBranchStub.resolves(branchInfo);
@@ -197,7 +188,7 @@ describe("apphosting rollouts", () => {
         pollOperationStub.onFirstCall().resolves(rollout);
         pollOperationStub.onSecondCall().resolves(build);
 
-        await createRollout(backendId, projectId, location, undefined, undefined, false);
+        await createRollout(backendId, projectId, undefined, undefined, false);
 
         expect(promptGitHubBranchStub).to.be.called;
         expect(createBuildStub).to.be.called;
@@ -206,33 +197,31 @@ describe("apphosting rollouts", () => {
       });
 
       it("should throw an error if GitHub branch is not found", async () => {
-        getBackendForLocationStub.resolves(backend);
-        getBackendForAmbiguousLocationStub.resolves(backend);
+        getBackend.resolves(backend);
         getRepoDetailsFromBackendStub.resolves(repoLinkDetails);
         listAllBranchesStub.resolves(branches);
 
         await expect(
-          createRollout(backendId, projectId, location, "invalid-branch", undefined, true),
+          createRollout(backendId, projectId, "invalid-branch", undefined, true),
         ).to.be.rejectedWith(/Unrecognized git branch/);
       });
 
       it("should throw an error if GitHub commit is not found", async () => {
-        getBackendForLocationStub.resolves(backend);
-        getBackendForAmbiguousLocationStub.resolves(backend);
+        getBackend.resolves(backend);
         getRepoDetailsFromBackendStub.resolves(repoLinkDetails);
         getGitHubCommitStub.rejects(new FirebaseError("error", { status: 422 }));
 
         await expect(
-          createRollout(backendId, projectId, location, undefined, commitSha, true),
+          createRollout(backendId, projectId, undefined, commitSha, true),
         ).to.be.rejectedWith(/Unrecognized git commit/);
       });
 
       it("should throw an error if --force flag is specified but --git-branch and --git-commit are missing", async () => {
-        getBackendForLocationStub.resolves(backend);
+        getBackend.resolves(backend);
         getRepoDetailsFromBackendStub.resolves(repoLinkDetails);
 
         await expect(
-          createRollout(backendId, projectId, location, undefined, undefined, true),
+          createRollout(backendId, projectId, undefined, undefined, true),
         ).to.be.rejectedWith(/Failed to create rollout with --force option/);
       });
     });