+++ src/apphosting/rollout.ts
@@ -13,7 +13,7 @@ import * as poller from "../operation-poller";
 import { logBullet, sleep } from "../utils";
 import { apphostingOrigin, consoleOrigin } from "../api";
 import { DeepOmit } from "../metaprogramming";
-import { getBackendForAmbiguousLocation, getBackendForLocation } from "./backend";
+import { getBackend } from "./backend";
 
 const apphostingPollerOptions: Omit<poller.OperationPollerOptions, "operationResourceName"> = {
   apiOrigin: apphostingOrigin(),
@@ -31,29 +31,19 @@ const GIT_COMMIT_SHA_REGEX = /^(?:[0-9a-f]{40}|[0-9a-f]{7})$/;
 export async function createRollout(
   backendId: string,
   projectId: string,
-  location: string,
   branch?: string,
   commit?: string,
   force?: boolean,
 ): Promise<void> {
-  let backend: apphosting.Backend;
-  if (location === "-" || location === "") {
-    backend = await getBackendForAmbiguousLocation(
-      projectId,
-      backendId,
-      "Please select the location of the backend you'd like to roll out:",
-      force,
-    );
-    location = apphosting.parseBackendName(backend.name).location;
-  } else {
-    backend = await getBackendForLocation(projectId, location, backendId);
-  }
+  const backend = await getBackend(projectId, backendId);
 
   if (!backend.codebase.repository) {
     throw new FirebaseError(
       `Backend ${backendId} is misconfigured due to missing a connected repository. You can delete and recreate your backend using 'firebase apphosting:backends:delete' and 'firebase apphosting:backends:create'.`,
     );
   }
+
+  const { location } = apphosting.parseBackendName(backend.name);
   const { repoLink, owner, repo, readToken } = await getRepoDetailsFromBackend(
     projectId,
     location,