+++ src/apphosting/secrets/dialogs.ts
@@ -207,8 +207,14 @@ function toUpperSnakeCase(key: string): string {
     .toUpperCase();
 }
 
-export async function envVarForSecret(secret: string): Promise<string> {
-  const upper = toUpperSnakeCase(secret);
+export async function envVarForSecret(
+  secret: string,
+  trimTestPrefix: boolean = false,
+): Promise<string> {
+  let upper = toUpperSnakeCase(secret);
+  if (trimTestPrefix && upper.startsWith("TEST_")) {
+    upper = upper.substring("TEST_".length);
+  }
   if (upper === secret) {
     try {
       env.validateKey(secret);