+++ firebase-vscode/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [Added] Added rerun execution button in variables context
 - [Added] Provide default required variables during execution
+- [Fixed] Fixed an issue where nvironment variables provided in `extraEnv` were not respected in some cases
 
 ## 0.14.2